% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{barcode_sample_type}
\alias{barcode_sample_type}
\title{Extracts the sample type (tumor/normal) from the barcode}
\usage{
barcode_sample_type(x, ...)
}
\arguments{
\item{x}{a vector of TCGA barcodes}
}
\description{
The sample type is encoded in the TCGA barcodes at the \code{[]} location,
\code{TCGA-02-0001-[01]C-01D-0183-01}. Codes between 01-09 are tumors, 10-19
are normal.
}
\details{
More detailed code information can be found in the following file:
\code{inst/extdata/tcga-barcode-codebook.xlsx,sample_type}
}
