% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gene_expression.R
\name{load_gene_info}
\alias{load_gene_info}
\title{Load the gene-level information}
\usage{
load_gene_info(xena.dir, ...)
}
\description{
You need to have run the \code{inst/scripts/extract-gencode-info.R} script over
the gencode v23 files and save the outputs to \code{xena.dir}.
}
\details{
The gene level information will be returned in the order expected for the
FacileDataSet
}
