% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sample_covariates.R
\name{parse_base_sample_covariates}
\alias{parse_base_sample_covariates}
\title{Base sample covariates}
\usage{
parse_base_sample_covariates(xena.dir, ...)
}
\value{
a tibble of id,indication,sex,sample_type
}
\description{
Split sample IDs indications, and provide sex, sample_type, and
stage annotation
}
