% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{parse_tcga_barcode}
\alias{parse_tcga_barcode}
\title{Converts a barcode into a data.frame of semantic pieces.}
\usage{
parse_tcga_barcode(x, ...)
}
\arguments{
\item{x}{a character vector of TCGA barcodes}
}
\description{
Each bit of the TCGA barcode means something, as outlined here:
https://docs.gdc.cancer.gov/Encyclopedia/pages/TCGA_Barcode/
}
\details{
A full barcode follows this format: \code{TCGA-02-0001-[01]C-01D-0183-01}, however
the data downloaded from xena mostly looks like: \code{TCGA-02-0002-01}
``
}
\examples{
barcodes <- c("TCGA-CV-5976-11", "TCGA-EJ-5532-01", "TCGA-D3-A1Q7-06",
              "TCGA-06-0211-02", "TCGA-AB-2989-03R")
parsed <- parse_tcga_barcode(barcodes)
}
