% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gene_expression.R
\name{prep_main_gene_expression}
\alias{prep_main_gene_expression}
\title{Prepare gene expression as main datasource for dataset construction.}
\usage{
prep_main_gene_expression(xena.dir, ..., debug = FALSE)
}
\arguments{
\item{xena.dir}{The path to the directory where you've downloaded the xena
data files. This function will load the \code{tcga_gene_expected_count.gz} and
\code{tcga_gene_expected_count.json} files}
}
\value{
A list of DGELists with some $samples info, names are TCGA
indications
}
\description{
The estimated RSEM gene-level counts from \code{tcga_gene_expected_count.gz} will
be used as the main assay for the dataset. This means a number of things, but
one thing it means is that if we don't make an attempt to add the basel-level
covariates for samples not found in the \code{tcga_gene_expected_count.gz}, they
won't appear anywhere in the final FacileTcgaDataSet.
}
\details{
This function will parse this count file into a list of DGEList objects
(one per indication). The DGELists will have basic annotation tied to them,
as generated from the \code{\link[=parse_base_sample_covariates]{parse_base_sample_covariates()}} function.

To achieve this, we wrangle / harmonize three sources of data:
\enumerate{
\item The counts data (\code{tcga_gene_expected_count.gz})
\item The gene-level data (generated in \code{scripts/extract-gencode-info.R})
\item A base set of sample-level covariate data, generated in
\code{\link[=parse_base_sample_covariates]{parse_base_sample_covariates()}}.
}
}
